Move coreblocks utils to transactron
