+++ coreblocks/frontend/rvc.py
@@ -2,7 +2,7 @@ from amaranth import *
 
 from transactron import TModule
 from coreblocks.params import *
-from coreblocks.utils import ValueLike
+from transactron.utils import ValueLike
 
 
 # An instruction or an instruction with the valid signal
