+++ coreblocks/params/layouts.py
@@ -1,6 +1,6 @@
 from coreblocks.params import GenParams, OpType, Funct7, Funct3
 from coreblocks.params.isa import ExceptionCause
-from coreblocks.utils.utils import layout_subset
+from transactron.utils.utils import layout_subset
 
 __all__ = [
     "SchedulerLayouts",
