+++ coreblocks/structs_common/csr.py
@@ -3,7 +3,7 @@ from amaranth.lib.enum import IntEnum
 from dataclasses import dataclass
 
 from transactron import Method, def_method, Transaction, TModule
-from coreblocks.utils import assign, bits_from_int
+from transactron.utils import assign, bits_from_int
 from coreblocks.params.genparams import GenParams
 from coreblocks.params.dependencies import DependencyManager, ListKey
 from coreblocks.params.fu_params import BlockComponentParams
