+++ scripts/synthesize.py
@@ -13,7 +13,7 @@ if __name__ == "__main__":
     sys.path.insert(0, parent)
 
 
-from coreblocks.utils.utils import ModuleConnector
+from transactron.utils.utils import ModuleConnector
 from coreblocks.params.genparams import GenParams
 from coreblocks.params.fu_params import FunctionalComponentParams
 from coreblocks.core import Core
