+++ test/common/testbenchio.py
@@ -4,7 +4,7 @@ from typing import Optional, Callable
 from transactron.lib import AdapterBase
 from transactron.core import ValueLike, SignalBundle
 from transactron._utils import mock_def_helper
-from coreblocks.utils._typing import RecordIntDictRet, RecordValueDict, RecordIntDict
+from transactron.utils._typing import RecordIntDictRet, RecordValueDict, RecordIntDict
 from .functions import set_inputs, get_outputs, TestGen
