+++ test/frontend/test_fetch.py
@@ -11,7 +11,7 @@ from coreblocks.frontend.fetch import Fetch, UnalignedFetch
 from coreblocks.frontend.icache import ICacheInterface
 from coreblocks.params import *
 from coreblocks.params.configurations import test_core_config
-from coreblocks.utils import ModuleConnector
+from transactron.utils import ModuleConnector
 from ..common import TestCaseWithSimulator, TestbenchIO, def_method_mock
