+++ test/structs_common/test_exception.py
@@ -6,7 +6,7 @@ from coreblocks.params import GenParams
 from coreblocks.params.isa import ExceptionCause
 from coreblocks.params.configurations import test_core_config
 from transactron.lib import Adapter
-from coreblocks.utils.utils import ModuleConnector
+from transactron.utils.utils import ModuleConnector
 
 from ..common import *
